Implemented IODrivers class as per #75 (still need to move some functionality around)
